refactor runtime to exclude ts from package
