+++ packages/parser/package.json
@@ -42,8 +42,5 @@
     "@types/chalk": "^2.2.0",
     "@types/debug": "^4.1.4",
     "antlr4ts-cli": "0.5.0-alpha.4"
-  },
-  "peerDependencies": {
-    "typescript": "3.1 - 4"
   }
 }
