+++ packages/parser/src/index.ts
@@ -1,14 +1,14 @@
 export {
-  default as parseTypeScriptFile,
-  parseTSQuery,
-  TSQueryAST,
-} from './loader/typescript/index.js';
+  default as parseTSQuery,
+  Query as TSQueryAST,
+} from './loader/typescript/query.js';
 
 export { Param, ParamKey, ParamType } from './loader/typescript/query.js';
 
 export {
   default as parseSQLFile,
   SQLQueryAST,
+  ParseEvent,
   SQLQueryIR,
   prettyPrintEvents,
   queryASTToIR,
